+++ examples/asr/experimental/conformer/conformer_linear_decoder_bpe.yaml
@@ -121,10 +121,6 @@ model:
       min_lr: 1e-6
       last_epoch: -1
 
-    var_noise:
-      std: 0.0 # variance of the Gaussian noise, recommended: 0.075
-      start_step: 0
-
 trainer:
   gpus: 0 # number of gpus
   num_nodes: 1
