+++ nemo/core/classes/modelPT.py
@@ -793,14 +793,6 @@ class ModelPT(LightningModule, Model):
         if lr is not None:
             optimizer_args['lr'] = lr
 
-        if 'var_noise' in optimizer_args:
-            self._var_noise_std = optimizer_args['var_noise'].get('std', None)
-            self._var_noise_start = optimizer_args['var_noise'].get('start_step', 0)
-            optimizer_args.pop('var_noise')
-        else:
-            self._var_noise_std = 0
-            self._var_noise_start = 0
-
         # Actually instantiate the optimizer
         if optimizer_cls is not None:
             if inspect.isclass(optimizer_cls):
@@ -1281,14 +1273,3 @@ class ModelPT(LightningModule, Model):
     def use_eff_save() -> bool:
         global _MODEL_EFF_SAVE
         return _MODEL_EFF_SAVE
-
-    def on_after_backward(self):
-        super().on_after_backward()
-
-        if self._var_noise_std > 0 and self.global_step >= self._var_noise_start:
-            for param_name, param in self.named_parameters():
-                if param.grad is not None:
-                    noise = torch.normal(
-                        mean=0.0, std=self._var_noise_std, size=param.size(), device=param.device, dtype=param.dtype
-                    )
-                    param.grad.data.add_(noise)
